requirements.txt
crawl4ai==0.7.4
fastapi==0.117.1
playwright==1.55.0
pydantic==2.11.9
pysnowflake==0.1.3
Requests==2.32.5
SQLAlchemy==2.0.43
sqlmodel==0.0.25
uvicorn==0.37.0
